+++ src/rendering/Rendering.h
@@ -46,7 +46,7 @@ public:
     void render();
 
     /// Set the rendering viewport (left, bottom, width, height) in device pixel units
-    void setDeviceViewport( const glm::vec4& deviceViewport );
+    void setDeviceViewport( const glm::ivec4& deviceViewport );
 
     /// Update all texture interpolation parameters for the active image component
     void updateImageInterpolation( const uuids::uuid& imageUid );
